REFAC: adjust the file locations of the "travt" module
